requirements.txt
numpy==1.20.3
pandas==1.3.4
Pillow==9.3.0
scipy==1.7.1
torchmetrics>=0.10.2
ffcv==0.0.3rc1
transformers==4.39.1
datasets==2.16.1
